services/docs-core/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "angineer-docs-core"
version = "0.1.0"
description = "AnGIneer docs engine"
requires-python = ">=3.10"
dependencies = [
    "mineru-rag[rag]>=0.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "chromadb>=0.5.5",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
